+++ .gitignore
@@ -9,3 +9,4 @@ __pycache__/
 gcp_credentials.json
 sample.env
 extraction.py
+docker-compose.yml
