+++ io_processing.py
@@ -1,15 +1,20 @@
-from logger import logger
-from translator import *
 import time
+
+from logger import logger
+from translation.api import translator
+
 
 
 
 def process_incoming_voice(file_url, input_language):
+    """
+    Main Function for processing audio based queries
+    """
     error_message = None
     try:
-        regional_text = audio_input_to_text(file_url, input_language)
+        regional_text = translator.speech_to_text(file_url, input_language)
         try:
-            english_text = indic_translation(text=regional_text, source=input_language, destination='en')
+            english_text = translator.translate_text(text=regional_text, source=input_language, destination='en')
         except Exception as e:
             error_message = "Indic translation to English failed"
             logger.error(f"Exception occurred: {e}", exc_info=True)
@@ -23,9 +28,12 @@ def process_incoming_voice(file_url, input_language):
 
 
 def process_incoming_text(regional_text, input_language):
+    """
+    Main function for processing text queries
+    """
     error_message = None
     try:
-        english_text = indic_translation(text=regional_text, source=input_language, destination='en')
+        english_text = translator.translate_text(text=regional_text, source=input_language, destination='en')
     except Exception as e:
         error_message = "Indic translation to English failed"
         english_text = None
@@ -34,9 +42,12 @@ def process_incoming_text(regional_text, input_language):
 
 
 def process_outgoing_text(english_text, input_language):
+    """
+    Main func for generating text response
+    """
     error_message = None
     try:
-        regional_text = indic_translation(text=english_text, source='en', destination=input_language)
+        regional_text = translator.translate_text(text=english_text, source='en', destination=input_language)
     except Exception as e:
         error_message = "English translation to indic language failed"
         logger.error(f"Exception occurred: {e}", exc_info=True)
@@ -45,8 +56,11 @@ def process_outgoing_text(english_text, input_language):
 
 
 def process_outgoing_voice(message, input_language):
+    """
+    Main function for generating audio response
+    """
     error_message = None
-    decoded_audio_content = text_to_speech(language=input_language, text=message)
+    decoded_audio_content = translator.text_to_speech(language=input_language, text=message)
     if decoded_audio_content is not None:
         logger.info("Creating output MP3 file")
         time_stamp = time.strftime("%Y%m%d-%H%M%S")
