+++ requirements-prod.txt
@@ -13,6 +13,7 @@ email-validator==1.3.1
 PyPDF2==3.0.1
 pypdf==3.17.1
 pydub==0.25.1
+google-cloud-storage==2.14.0
 google-cloud-texttospeech==2.14.1
 google-cloud-translate==3.11.0
 google-cloud-speech==2.18.0
